Changed Saveable attribute Position property type to long as it should be

+++ SaveableAttribute.cs
@@ -12,6 +12,6 @@ namespace SaveableDotNet
         /// Saveable position in stream
         /// </summary>
         /// <value></value>
-        public int Position { get; set; } = -1;
+        public long Position { get; set; } = -1;
     }
 }
